* fixed new brightness_contrast_gamma function to work on 16bit data 	* introduced new getRGB/setRGB with double arguments for more correct 	  and comfortable handling 	* needed to explicitly cast some other code places as the with the 	  overloaded getRGB/setRGB the code became ambegious

+++ trunk/econvert/econvert.cc
@@ -769,7 +769,7 @@ bool convert_descew2 (const Argument<bool>& arg)
 	    
 	    true
 	    )
-	  it2.setRGB (255, 255, 0), it2.set (it2);
+	  it2.setRGB ((uint16_t)255, (uint16_t)255, (uint16_t)0), it2.set (it2);
 	
 	++it1; ++it2; ++it3; ++it4; ++it5;
       }
